pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "tabular_page_maker"
authors = [{name = "John C. G. Sturdy", email = "jcg.sturdy@gmail.com"}]
description = "Create tabular pages from YAML."
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)"]
version = "0.0.1"
dependencies = [
  "pyyaml"
]